Minor improvements for Windows.
* Fix Appveyor building.
* Squash casting warnings.

+++ src/LuaManager.h
@@ -300,7 +300,7 @@ inline double get_optional_double(lua_State* L, int index, char const* field, do
 	lua_getfield(L, index, field);
 	if(lua_isnumber(L, -1))
 	{
-		ret= lua_tonumber(L, -1);
+		ret= static_cast<double>(lua_tonumber(L, -1));
 	}
 	lua_pop(L, 1);
 	return ret;
@@ -312,7 +312,7 @@ inline int get_optional_int(lua_State* L, int index, char const* field, int def)
 	lua_getfield(L, index, field);
 	if(lua_isnumber(L, -1))
 	{
-		ret= lua_tonumber(L, -1);
+		ret= static_cast<int>(lua_tonumber(L, -1));
 	}
 	lua_pop(L, 1);
 	return ret;
